refactor: 优化在线用户模块和代码清理
- 移除未使用的 sa-token-jwt 依赖，改用 Redis 存储模式
- 在线用户模块：添加登录时间存储，移除会话ID，强制退出改用用户ID
- GlobalExceptionHandler 移动到 framework 包
- PageRequest 分页参数类型从 Integer 改为 Long
- JobServiceImpl 使用 page.convert() 简化分页转换
- Sa-Token 配置：禁止并发登录，每次生成新 token

+++ fast-serve/src/main/java/com/fast/common/result/PageRequest.java
@@ -18,21 +18,21 @@ public class PageRequest {
      * 页码（从 1 开始）
      */
     @Min(1)
-    private Integer pageNum = 1;
+    private Long pageNum = 1L;
 
     /**
      * 每页数量
      */
     @Min(1)
     @Max(100)
-    private Integer pageSize = 10;
+    private Long pageSize = 10L;
 
     /**
      * 获取起始偏移量（用于 SQL LIMIT offset）
      *
      * @return 偏移量
      */
-    public int getOffset() {
+    public long getOffset() {
         return (pageNum - 1) * pageSize;
     }
